# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LLM Eiskunstlauf Championship API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessMessage {
  message: string;
}

export type CreateGatewayType =
  (typeof CreateGatewayType)[keyof typeof CreateGatewayType];

export const CreateGatewayType = {
  openrouter: "openrouter",
  github_copilot: "github_copilot",
  custom: "custom",
} as const;

export interface CreateGateway {
  name: string;
  type: CreateGatewayType;
  baseUrl: string;
  apiKey: string;
}

export interface CreateDataset {
  name: string;
  content: string;
  systemPrompt: string;
}

export interface SessionSyncRequest {
  gateways?: CreateGateway[];
  datasets?: CreateDataset[];
}

export interface SessionSyncResponse {
  sessionId: string;
  gatewayCount: number;
  datasetCount: number;
}

export type GatewayType = (typeof GatewayType)[keyof typeof GatewayType];

export const GatewayType = {
  openrouter: "openrouter",
  github_copilot: "github_copilot",
  custom: "custom",
} as const;

export interface Gateway {
  id: number;
  name: string;
  type: GatewayType;
  baseUrl: string;
  createdAt: string;
}

export interface ModelInfo {
  id: string;
  name: string;
  gatewayId: number;
  gatewayName: string;
}

export type DatasetPrivacyStatus =
  (typeof DatasetPrivacyStatus)[keyof typeof DatasetPrivacyStatus];

export const DatasetPrivacyStatus = {
  unchecked: "unchecked",
  clean: "clean",
  issues_found: "issues_found",
  anonymized: "anonymized",
} as const;

export interface Dataset {
  id: number;
  name: string;
  content: string;
  systemPrompt: string;
  privacyStatus: DatasetPrivacyStatus;
  /** @nullable */
  privacyReport?: string | null;
  createdAt: string;
}

export interface PrivacyCheckRequest {
  gatewayId: number;
  modelId: string;
}

export type PrivacyCheckResultStatus =
  (typeof PrivacyCheckResultStatus)[keyof typeof PrivacyCheckResultStatus];

export const PrivacyCheckResultStatus = {
  clean: "clean",
  issues_found: "issues_found",
} as const;

export interface PrivacyFinding {
  type: string;
  text: string;
  suggestion: string;
}

export interface PrivacyCheckResult {
  status: PrivacyCheckResultStatus;
  findings: PrivacyFinding[];
  report: string;
}

export interface AnonymizeRequest {
  gatewayId: number;
  modelId: string;
}

export interface GenerateDatasetRequest {
  name: string;
  topic: string;
  systemPrompt: string;
  numberOfItems: number;
  gatewayId: number;
  modelId: string;
}

export type CompetitionStatus =
  (typeof CompetitionStatus)[keyof typeof CompetitionStatus];

export const CompetitionStatus = {
  draft: "draft",
  running: "running",
  completed: "completed",
  error: "error",
} as const;

export interface Competition {
  id: number;
  name: string;
  datasetId: number;
  status: CompetitionStatus;
  createdAt: string;
}

export interface ModelSelection {
  gatewayId: number;
  modelId: string;
  modelName: string;
}

export interface CreateCompetition {
  name: string;
  datasetId: number;
  systemPrompt: string;
  contestantModels: ModelSelection[];
  judgeModels: ModelSelection[];
}

export type CompetitionDetailStatus =
  (typeof CompetitionDetailStatus)[keyof typeof CompetitionDetailStatus];

export const CompetitionDetailStatus = {
  draft: "draft",
  running: "running",
  completed: "completed",
  error: "error",
} as const;

export interface JudgeScore {
  judgeModelId: string;
  judgeModelName: string;
  score: number;
  reasoning: string;
}

export interface ModelResponse {
  dataItemIndex: number;
  response: string;
  durationMs: number;
  promptTokens: number;
  completionTokens: number;
  cost: number;
  judgeScores: JudgeScore[];
}

export interface CompetitionResult {
  gatewayId: number;
  modelId: string;
  modelName: string;
  avgSpeed: number;
  avgCost: number;
  avgQuality: number;
  totalTokens: number;
  responses: ModelResponse[];
}

export interface CompetitionDetail {
  id: number;
  name: string;
  datasetId: number;
  systemPrompt: string;
  status: CompetitionDetailStatus;
  contestantModels: ModelSelection[];
  judgeModels: ModelSelection[];
  results: CompetitionResult[];
  createdAt: string;
}

export type UploadDatasetBody = {
  /** A .md (Markdown) file */
  file: Blob;
  /** Dataset name (defaults to filename if omitted) */
  name?: string;
  /** System prompt for evaluating responses */
  systemPrompt: string;
};