# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LLM Eiskunstlauf Championship API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /datasets/{id}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get a single dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dataset detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
    delete:
      operationId: deleteDataset
      tags: [datasets]
      summary: Delete a dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
...
  /competitions:
    get:
      operationId: listCompetitions
      tags: [competitions]
      summary: List all competitions
      responses:
        "200":
          description: List of competitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competition"
    post:
      operationId: createCompetition
      tags: [competitions]
      summary: Create a new competition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompetition"
      responses:
        "201":
          description: Competition created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competition"
  /competitions/{id}:
    get:
      operationId: getCompetition
      tags: [competitions]
      summary: Get competition details with results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competition with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDetail"
    delete:
      operationId: deleteCompetition
      tags: [competitions]
      summary: Delete a competition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /competitions/{id}/run:
    post:
      operationId: runCompetition
      tags: [competitions]
      summary: Start running a competition (evaluate all models asynchronously)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competition run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDetail"
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Dataset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        content:
          type: string
        systemPrompt:
          type: string
        privacyStatus:
          type: string
          enum: [unchecked, clean, issues_found, anonymized]
        privacyReport:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - content
        - systemPrompt
        - privacyStatus
        - createdAt
...
    Competition:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        datasetId:
          type: integer
        status:
          type: string
          enum: [draft, running, completed, error]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - datasetId
        - status
        - createdAt
    CreateCompetition:
      type: object
      properties:
        name:
          type: string
        datasetId:
          type: integer
        systemPrompt:
          type: string
        contestantModels:
          type: array
          items:
            $ref: "#/components/schemas/ModelSelection"
        judgeModels:
          type: array
          items:
            $ref: "#/components/schemas/ModelSelection"
      required:
        - name
        - datasetId
        - systemPrompt
        - contestantModels
        - judgeModels
    ModelSelection:
      type: object
      properties:
        gatewayId:
          type: integer
        modelId:
          type: string
        modelName:
          type: string
      required:
        - gatewayId
        - modelId
        - modelName
    CompetitionDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        datasetId:
          type: integer
        systemPrompt:
          type: string
        status:
          type: string
          enum: [draft, running, completed, error]
        contestantModels:
          type: array
          items:
            $ref: "#/components/schemas/ModelSelection"
        judgeModels:
          type: array
          items:
            $ref: "#/components/schemas/ModelSelection"
        results:
          type: array
          items:
            $ref: "#/components/schemas/CompetitionResult"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - datasetId
        - systemPrompt
        - status
        - contestantModels
        - judgeModels
        - results
        - createdAt
    CompetitionResult:
      type: object
      properties:
        gatewayId:
          type: integer
        modelId:
          type: string
        modelName:
          type: string
        avgSpeed:
          type: number
        avgCost:
          type: number
        avgQuality:
          type: number
        totalTokens:
          type: integer
        responses:
          type: array
          items:
            $ref: "#/components/schemas/ModelResponse"
      required:
        - gatewayId
        - modelId
        - modelName
        - avgSpeed
        - avgCost
        - avgQuality
        - totalTokens
        - responses
    ModelResponse:
      type: object
      properties:
        dataItemIndex:
          type: integer
        response:
          type: string
        durationMs:
          type: number
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        cost:
          type: number
        judgeScores:
          type: array
          items:
            $ref: "#/components/schemas/JudgeScore"
      required:
        - dataItemIndex
        - response
        - durationMs
        - promptTokens
        - completionTokens
        - cost
        - judgeScores
    JudgeScore:
      type: object
      properties:
        judgeModelId:
          type: string
        judgeModelName:
          type: string
        score:
          type: integer
        reasoning:
          type: string
      required:
        - judgeModelId
        - judgeModelName
        - score
        - reasoning